Final game
This is my final game. The project is called FINAL GAME. Everything
works fine, but you may need to run it a couple times seeing that
sometimes the enemy and character collide right away. The object of the
game is to run away from the enemies and get the highest score possible.